streamlit-app/requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
google-cloud-bigquery>=3.0.0
google-cloud-storage>=2.0.0
